+++ src/Encamina.Enmarcha.Bot.Skills.QuestionAnswering/QuestionAnsweringDialogServices.cs
@@ -9,10 +9,10 @@ using Microsoft.Bot.Builder;
 namespace Encamina.Enmarcha.Bot.Skills.QuestionAnswering;
 
 internal record class QuestionAnsweringDialogServices(ICognitiveServiceFactory<IQuestionAnsweringService> QuestionAnsweringCognitiveServiceFactory,
-    IBotTelemetryClient BotTelemetryClient = null,
-    ISourcesProcessor SourcesProcessor = null,
-    IMetadataProcessor MetadataProcessor = null,
-    IQuestionRequestProcessor QuestionRequestProcessor = null,
-    IQuestionResultProcessor QuestionResultProcessor = null,
-    ISendAnswersProcessor SendAnswersProcessor = null,
-    IIntentResponsesProvider IntentResponsesProvider = null);
+    IBotTelemetryClient? BotTelemetryClient = null,
+    ISourcesProcessor? SourcesProcessor = null,
+    IMetadataProcessor? MetadataProcessor = null,
+    IQuestionRequestProcessor? QuestionRequestProcessor = null,
+    IQuestionResultProcessor? QuestionResultProcessor = null,
+    ISendAnswersProcessor? SendAnswersProcessor = null,
+    IIntentResponsesProvider? IntentResponsesProvider = null);